requirements.txt
absl-py==1.4.0
attrs==22.2.0
comtypes==1.1.14
contourpy==1.0.7
cv2-brightness==0.0.1
cycler==0.11.0
enum34==1.1.10
flatbuffers==23.3.3
fonttools==4.39.3
future==0.18.3
keyboard==0.13.5
kiwisolver==1.4.4
matplotlib==3.7.1
mediapipe==0.9.2.1
MouseInfo==0.1.3
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.5.0
protobuf==4.22.1
psutil==5.9.4
PyAutoGUI==0.9.53
pycaw==20230322
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pywin32==306
screen-brightness-control==0.20.0
six==1.16.0
WMI==1.5.1
